refactor: extract EditUserAvatarContent as a mirror of EditAgentAvatarContent
